测试分支

+++ micro-service/service-B/src/main/java/com/zhang/study/controller/FileUploadController.java
@@ -1,17 +1,21 @@
 package com.zhang.study.controller;
 
+import com.zhang.out.entity.FileSyncTransPart;
+import com.zhang.out.mapper.FileSyncTransMapper;
 import com.zhang.study.base.Result;
 import com.zhang.study.entity.TreeDemo;
+import com.zhang.utils.MD5Utils;
+import com.zhang.utils.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.annotation.Resource;
 import java.io.*;
 import java.net.URLDecoder;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.concurrent.*;
 
 
 /**
@@ -28,6 +32,9 @@ public class FileUploadController {
     @Value("${saveFilePath}")
     private String saveFilePath;
 
+    @Resource
+    private FileSyncTransMapper fileSyncTransMapper;
+
     /**
      * 简单的单文件上传
      *
@@ -36,7 +43,7 @@ public class FileUploadController {
      * @throws IOException
      */
     @RequestMapping(value = "/upload", method = RequestMethod.POST)
-    public Map<String, String> upload(@RequestParam(value = "file", required = false) MultipartFile file, String guid, String chunkSize, String fileName) throws IOException {
+    public Result upload(@RequestParam(value = "file", required = false) MultipartFile file, String guid, String chunkSize, String fileName) throws IOException {
         guid = URLDecoder.decode(guid, "UTF-8");
 
         BufferedInputStream inputStream = null;
@@ -60,29 +67,27 @@ public class FileUploadController {
             outs.close();
             bouts.close();
         } catch (Exception e) {
-            throw new IOException(e.getMessage());
-        } finally {
-        }
-        try {
-            if (null != inputStream) {
-                inputStream.close();
-            }
-            if (null != outs) {
-                outs.close();
-            }
-            if (null != bouts) {
-                bouts.close();
-            }
-        } catch (Exception e) {
-            throw new IOException(e.getMessage());
-        }
+            e.printStackTrace();
+            return Result.error("上传失败");
 
-        Map<String, String> map = new HashMap<String, String>();
-        String result = "";
-        int res = -1;
+        } finally {
+            try {
+                if (null != inputStream) {
+                    inputStream.close();
+                }
+                if (null != outs) {
+                    outs.close();
+                }
+                if (null != bouts) {
+                    bouts.close();
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                return Result.error("上传失败");
+            }
+        }
         //返回提示信息
-        map.put("result", result);
-        return map;
+        return Result.ok("上传成功");
     }
 
 
@@ -94,18 +99,18 @@ public class FileUploadController {
      * @throws IOException
      */
     @RequestMapping(value = "/uploadNoChunk", method = RequestMethod.POST)
-    public Map<String, String> uploadNoChunk(@RequestParam(value = "file", required = false) MultipartFile file, String guid, String filePath) throws IOException {
+    public Result uploadNoChunk(@RequestParam(value = "file", required = false) MultipartFile file, String guid, String filePath, String fileName) throws IOException {
         String result = "";
         guid = URLDecoder.decode(guid, "UTF-8");
         BufferedInputStream inputStream = null;
         OutputStream outs = null;
         BufferedOutputStream bouts = null;
         try {
-            File file1 = new File(saveFilePath + filePath.substring(0,filePath.lastIndexOf("/")));
-            if(!file1.exists()){
+            File file1 = new File(saveFilePath + filePath.substring(0, filePath.lastIndexOf("/")));
+            if (!file1.exists()) {
                 file1.mkdirs();
             }
-            File dirFile = new File(saveFilePath + filePath.substring(0,filePath.lastIndexOf("/")), file.getOriginalFilename());
+            File dirFile = new File(saveFilePath + filePath.substring(0, filePath.lastIndexOf("/")), fileName);
             //以读写的方式打开目标文件
             inputStream = new BufferedInputStream(file.getInputStream());
             outs = new FileOutputStream(dirFile);
@@ -121,8 +126,10 @@ public class FileUploadController {
             inputStream.close();
             outs.close();
             bouts.close();
+//            saveDataSouce(saveFilePath + filePath.substring(0, filePath.lastIndexOf("/")) + "/" + file.getOriginalFilename());
         } catch (Exception e) {
-            throw new IOException(e.getMessage());
+            e.printStackTrace();
+            return Result.error("上传失败");
         } finally {
         }
         try {
@@ -136,15 +143,10 @@ public class FileUploadController {
                 bouts.close();
             }
         } catch (Exception e) {
-            throw new IOException(e.getMessage());
+            e.printStackTrace();
+            return Result.error("上传失败");
         }
-
-        Map<String, String> map = new HashMap<String, String>();
-
-        int res = -1;
-        //返回提示信息
-        map.put("result", result);
-        return map;
+        return Result.ok("上传成功");
     }
 
 
@@ -178,6 +180,7 @@ public class FileUploadController {
 
     /**
      * 合并文件
+     * 在拆分存放
      */
     @RequestMapping(value = "mergeFile", method = RequestMethod.POST)
     @ResponseBody
@@ -187,7 +190,7 @@ public class FileUploadController {
         OutputStream outs = null;
         BufferedOutputStream bouts = null;
         File[] listfile = null;
-
+        Boolean check = true;
         try {
             File file = new File(checkMD5Path + treeDemo.getGuid());
             if (file.exists()) {
@@ -197,13 +200,17 @@ public class FileUploadController {
             if (!file1.exists()) {
                 file1.mkdirs();
             }
-            //申明随机读取文件RandomAccessFile
             outs = new FileOutputStream(new File(saveFilePath + treeDemo.getPathUrl()));
             bouts = new BufferedOutputStream(outs);
             //开始合并文件，对应切片的二进制文件
             for (int i = 0; i < listfile.length; i++) {
+                if (i == 99) {
+                    System.out.println(i);
+                }
                 //读取切片文件
-                ins = new FileInputStream(checkMD5Path + treeDemo.getGuid() + "\\" + treeDemo.getName() + "_" + i * 5 * 1024 * 1024);
+                String pathUrl = checkMD5Path + treeDemo.getGuid() + "\\" + treeDemo.getName() + "_" + i;
+                System.out.println(pathUrl);
+                ins = new FileInputStream(pathUrl);
                 bins = new BufferedInputStream(ins);
                 byte[] buff = new byte[8192];
                 int n = 0;
@@ -214,14 +221,20 @@ public class FileUploadController {
                 ins.close();
                 bins.close();
             }
+            //
             bouts.flush();// 这里一定要调用flush()方法
             outs.flush();
-//            ins.close();
-//            bins.close();
             outs.close();
             bouts.close();
+            System.out.println("合并完成");
+//-------------------------------------------------------------------------------------------------//
+//            check = saveDataSouce(saveFilePath + treeDemo.getPathUrl());
+//-------------------------------------------------------------------------------------------------//
+
+
         } catch (Exception e) {
             e.printStackTrace();
+            return Result.error("上传失败!");
         } finally {
             try {
                 if (null != ins) {
@@ -240,9 +253,95 @@ public class FileUploadController {
                 e.printStackTrace();
             }
         }
-        return Result.ok("上传成功!");
+        if (check) {
+            return Result.ok("上传成功!");
+
+        } else {
+            return Result.error("上传失败!");
+
+        }
     }
 
+
+    /**
+     * 设置文件流md5校验,并重新分片
+     */
+    public Boolean saveDataSouce(String filePath) throws Exception {
+        InputStream is = null;
+        BufferedInputStream bis = null;
+        File file = new File(filePath);
+        String fileId = "";
+        Boolean check = true;
+        try {
+            fileId = MD5Utils.getFileMD5String(file);
+            //查询库中是否存在,存在则不需要录入
+            int count = fileSyncTransMapper.selectFileIdCount(fileId);
+            if (count == 0) {
+                is = new FileInputStream(file);
+                long fileSize = is.available();
+                // 分割大小500M
+                long splitSize = 500 * 1024 * 1024;
+                // 计算分割数量
+                int fileTotal = 0;
+                // 判断大小
+                if (fileSize > splitSize) {
+                    if (fileSize % splitSize == 0) {
+                        fileTotal = (int) (fileSize / splitSize);
+                    } else {
+                        fileTotal = (int) (fileSize / splitSize) + 1;
+                    }
+                } else {
+                    fileTotal = 1;
+                    splitSize = fileSize;
+                }
+                // 创建字节数组
+                byte[] bytes;
+                bis = new BufferedInputStream(is);
+                for (int i = 0; i < fileTotal; i++) {
+                    long availableSize = is.available();
+                    if (availableSize < splitSize) {
+                        bytes = new byte[Integer.parseInt(String.valueOf(availableSize).substring(i * (int) splitSize, (int) splitSize))];
+                    } else {
+                        bytes = new byte[(int) splitSize];
+                    }
+                    int len = bis.read(bytes);
+                    //循环保存3次，容错
+//                    check = saveFileTransPart(fileId, bytes, i);
+//                    重新分片文件
+
+                    if (!check) {
+                        break;
+                    }
+                }
+            }
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (null != is) {
+                try {
+                    is.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return check;
+    }
+
+    private boolean saveFileTransPart(String fileId, byte[] bytes, int i) {
+        FileSyncTransPart part = new FileSyncTransPart();
+        part.setFileId(fileId);
+        part.setPartFile(bytes);
+        part.setPartSort(i);
+        try {
+            fileSyncTransMapper.saveFileTransPart(part);
+        } catch (Exception e) {
+            return false;
+        }
+        return true;
+    }
 
     /**
      * 删除缓存文件
@@ -255,7 +354,10 @@ public class FileUploadController {
     public Result delList(@RequestBody List<TreeDemo> treeDemoList) {
         if (!CollectionUtils.isEmpty(treeDemoList)) {
             for (TreeDemo treeDemo : treeDemoList) {
-                deleteDir(checkMD5Path + treeDemo.getGuid());
+                File file = new File(checkMD5Path + treeDemo.getGuid());
+                if (file.exists()) {
+                    deleteDir(checkMD5Path + treeDemo.getGuid());
+                }
             }
         }
         return Result.ok("删除成功");
@@ -278,5 +380,76 @@ public class FileUploadController {
             }
         }
     }
+
+
+    @RequestMapping(value = "tree", method = RequestMethod.GET)
+    @ResponseBody
+    public Result treadCallable() {
+        ExecutorService executorService = Executors.newFixedThreadPool(10);
+        Map<String, Future<String>> uploadMap = new HashMap<>();
+        //存放异常的数据
+        Map<String, String> resultMap = new HashMap<>();
+        List<String> exceptionList = new ArrayList<>();
+        for (int i = 0; i < 11; i++) {
+            Callable callable1 = new CallableImpl(i);
+            Future<String> future = executorService.submit(callable1);
+            uploadMap.put(String.valueOf(i), future);
+        }
+        uploadMap.forEach((key, value) -> {
+            Future<String> future = value;
+            try {
+                resultMap.put(key, future.get(1, TimeUnit.MINUTES));
+            } catch (InterruptedException e) {
+                exceptionList.add(key);
+                e.printStackTrace();
+            } catch (ExecutionException e) {
+                exceptionList.add(key);
+                e.printStackTrace();
+            } catch (TimeoutException e) {
+                exceptionList.add(key);
+                e.printStackTrace();
+            }
+
+        });
+
+        executorService.shutdown();
+        //对异常数据重新请求三次
+        if (CollectionUtils.isEmpty(exceptionList)) {
+            System.out.println("多线程结束!");
+            return Result.ok(resultMap);
+        } else {
+            return Result.error("失败");
+        }
+
+
+    }
+
+
+    public class CallableImpl implements Callable {
+        private int count;
+
+        CallableImpl(int count) {
+            this.count = count;
+        }
+
+        @Override
+        public Object call() {
+            try {
+                Thread.sleep(3000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            String uuid = uuid();
+            System.out.println("count>>>>>>>>>>>>>>>>>>>>>>>" + count);
+            System.out.println("count>>>>>>>>>>>>>>>>>>>>>>>" + uuid);
+            return uuid;
+        }
+    }
+
+    public String uuid() {
+
+        return StringUtils.uuid32len();
+    }
+
 
 }
